Clean up output file handling
Use a `TemporaryDirectory` object and create the output file in that directory. This means we don't have to worry about Windows and reading from a NamedTemporaryFile, and can guarantee that the directory is cleaned up even if there was an exception.

Then use `shutil.copyfileobj()` for a fast data transfer instead of shelling out to `cat` or `type`.

The faked file can be an in-memory `StringIO()` object, since nothing writes to it anyway.